backend/requirement.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
mysqlclient
requests
django-rest-framework-nested
drf-nested-routers
